Add files via upload
Predicting the percentage of a Student based on the number of Study Hours
The first task in Data Science and Business Analytics Internship at The Sparks Foundation
In this project we will use linear regression to predict a student grades based on the number of hours he studied. It involves only two variables.